Add easycheck and perftester
Assertions should be avoided in production code, and hence easycheck was proposed. It offers a number of functions that allow for checking various conditions when one would feel like using the assert statement. It also offers aliases to be used in testing; e.g., `check_if` (`assert_if`), `check_if_not` (`assert_if_not`), `check_if_isclose` (`assert_if_isclose`), `check_type` (`assert_type`), and the like.
The `perftester` package aims to allow for performance testing of Python functions; performance here means both execution time and memory usage. The package can be used inside `pytest`, but mainly as a separate testing framework. As far as I know, it is the only performance testing framework for Python functions. It also allows for benchmarking functions, in terms of time and memory. (Benchmarking is included to enable the user to define limits for performance tests.)